Added version 1.14.1077

+++ chromium/components/ntp_snippets/remote/remote_suggestions_database.cc
@@ -8,11 +8,12 @@
 
 #include "base/files/file_path.h"
 #include "base/memory/ptr_util.h"
+#include "base/sys_info.h"
 #include "base/task_scheduler/post_task.h"
 #include "components/leveldb_proto/proto_database_impl.h"
 #include "components/ntp_snippets/remote/proto/ntp_snippets.pb.h"
 
-using leveldb_env::SharedReadCache;
+using leveldb_proto::ProtoDatabase;
 using leveldb_proto::ProtoDatabaseImpl;
 
 namespace {
@@ -27,36 +28,50 @@ const char kSnippetDatabaseFolder[] = "snippets";
 const char kImageDatabaseFolder[] = "images";
 
 const size_t kDatabaseWriteBufferSizeBytes = 512 << 10;
+const size_t kDatabaseWriteBufferSizeBytesForLowEndDevice = 128 << 10;
 }  // namespace
 
 namespace ntp_snippets {
 
 RemoteSuggestionsDatabase::RemoteSuggestionsDatabase(
     const base::FilePath& database_dir)
-    : database_initialized_(false),
+    : RemoteSuggestionsDatabase(
+          base::CreateSequencedTaskRunnerWithTraits(
+              {base::MayBlock(), base::TaskPriority::BACKGROUND,
+               base::TaskShutdownBehavior::CONTINUE_ON_SHUTDOWN}),
+          database_dir) {}
+
+RemoteSuggestionsDatabase::RemoteSuggestionsDatabase(
+    scoped_refptr<base::SequencedTaskRunner> task_runner,
+    const base::FilePath& database_dir)
+    : RemoteSuggestionsDatabase(
+          std::make_unique<ProtoDatabaseImpl<SnippetProto>>(task_runner),
+          std::make_unique<ProtoDatabaseImpl<SnippetImageProto>>(task_runner),
+          database_dir) {}
+
+RemoteSuggestionsDatabase::RemoteSuggestionsDatabase(
+    std::unique_ptr<ProtoDatabase<SnippetProto>> database,
+    std::unique_ptr<ProtoDatabase<SnippetImageProto>> image_database,
+    const base::FilePath& database_dir)
+    : database_(std::move(database)),
+      database_initialized_(false),
+      image_database_(std::move(image_database)),
       image_database_initialized_(false),
       weak_ptr_factory_(this) {
-  auto file_task_runner = base::CreateSequencedTaskRunnerWithTraits(
-      {base::MayBlock(), base::TaskPriority::BACKGROUND,
-       base::TaskShutdownBehavior::CONTINUE_ON_SHUTDOWN});
-  database_ =
-      base::MakeUnique<ProtoDatabaseImpl<SnippetProto>>(file_task_runner);
-  image_database_ =
-      base::MakeUnique<ProtoDatabaseImpl<SnippetImageProto>>(file_task_runner);
-
   base::FilePath snippet_dir = database_dir.AppendASCII(kSnippetDatabaseFolder);
-  database_->InitWithOptions(
-      kDatabaseUMAClientName,
-      leveldb_proto::Options(snippet_dir, SharedReadCache::Default,
-                             kDatabaseWriteBufferSizeBytes),
-      base::Bind(&RemoteSuggestionsDatabase::OnDatabaseInited,
-                 weak_ptr_factory_.GetWeakPtr()));
+  leveldb_env::Options options = leveldb_proto::CreateSimpleOptions();
+  if (base::SysInfo::IsLowEndDevice()) {
+    options.write_buffer_size = kDatabaseWriteBufferSizeBytesForLowEndDevice;
+  } else {
+    options.write_buffer_size = kDatabaseWriteBufferSizeBytes;
+  }
+  database_->Init(kDatabaseUMAClientName, snippet_dir, options,
+                  base::Bind(&RemoteSuggestionsDatabase::OnDatabaseInited,
+                             weak_ptr_factory_.GetWeakPtr()));
 
   base::FilePath image_dir = database_dir.AppendASCII(kImageDatabaseFolder);
-  image_database_->InitWithOptions(
-      kImageDatabaseUMAClientName,
-      leveldb_proto::Options(image_dir, SharedReadCache::Default,
-                             kDatabaseWriteBufferSizeBytes),
+  image_database_->Init(
+      kImageDatabaseUMAClientName, image_dir, options,
       base::Bind(&RemoteSuggestionsDatabase::OnImageDatabaseInited,
                  weak_ptr_factory_.GetWeakPtr()));
 }
@@ -334,6 +349,8 @@ void RemoteSuggestionsDatabase::DeleteUnreferencedImages(
       keys_to_remove->emplace_back(key);
     }
   }
+  if (keys_to_remove->empty())
+    return;
   DeleteImages(std::move(keys_to_remove));
 }
